Separate LifeCycle from ProjectLifeCycle to allow LifeCycles to be shared across projects.
