--- db/migrate/20241101124648_add_life_cycle_to_work_packages.rb
@@ -1,5 +1,0 @@
-class AddLifeCycleToWorkPackages < ActiveRecord::Migration[7.1]
-  def change
-    add_reference :work_packages, :life_cycle, foreign_key: { to_table: :life_cycles }, null: true
-  end
-end
