+++ spec/models/project_life_cycle_spec.rb
@@ -1,6 +1,48 @@
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+
 require "rails_helper"
 
 RSpec.describe ProjectLifeCycle do
-  it { is_expected.to belong_to(:project).required(true) }
-  it { is_expected.to belong_to(:life_cycle).required(true) }
+  it "cannot be instantiated" do
+    expect { described_class.new }.to raise_error(NotImplementedError)
+  end
+
+  it "cannot be instantiated with an invalid type" do
+    expect { described_class.new(type: "InvalidType") }.to raise_error(ActiveRecord::SubclassNotFound)
+  end
+
+  it "can be instantiated with a valid type" do
+    expect { described_class.new(type: "Projects::Gate") }.not_to raise_error
+  end
+
+  # For more specs see:
+  # - spec/support/shared/project_life_cycle_helpers.rb
+  # - spec/models/projects/gate_spec.rb
+  # - spec/models/projects/stage_spec.rb
 end
