+++ spec/models/projects/life_cycle_spec.rb
@@ -28,21 +28,14 @@
 
 require "rails_helper"
 
-RSpec.describe Projects::LifeCycle do
-  it "cannot be instantiated" do
-    expect { described_class.new }.to raise_error(NotImplementedError)
+RSpec.describe LifeCycle do
+  describe "associations" do
+    it { is_expected.to have_many(:project_life_cycles).dependent(:destroy) }
+    it { is_expected.to have_many(:projects).through(:project_life_cycles) }
+    it { is_expected.to belong_to(:color).required(true) }
   end
 
-  it "cannot be instantiated with an invalid type" do
-    expect { described_class.new(type: "InvalidType") }.to raise_error(ActiveRecord::SubclassNotFound)
+  describe "validations" do
+    it { is_expected.to validate_presence_of(:name) }
   end
-
-  it "can be instantiated with a valid type" do
-    expect { described_class.new(type: "Projects::Gate") }.not_to raise_error
-  end
-
-  # For more specs see:
-  # - spec/support/shared/projects_life_cycle_helpers.rb
-  # - spec/models/projects/gate_spec.rb
-  # - spec/models/projects/stage_spec.rb
 end
